stop copying paths all the time
since create_pipe() needs to modify the path and its argument is read
only it would always copy the argument path.  That means it would copy
on every connection to the named pipe, instead we can make one copy in
bind() and use it in every call to CreateNamedPipeW().  Then we can
improve the way the buffer is generated by copying to_u16s() from the
standard library since the standard library copy isn't public.